tested_vesion_measurement_ok
this vesion was tested this morning with all variables set and get
functions working as well as the real measurement data

+++ Code ROV/Libraries/CMSIS/Include/arm_math.h
@@ -5387,7 +5387,7 @@ extern "C"
 	  float32_t *pYData = S->pYData;	    /* pointer to output table */
 
 	  /* Calculation of index */
-	  i =   (x - S->x1) / xSpacing;
+	  i = (int32_t)  ((x - S->x1) / xSpacing);
 
 	  if(i < 0)
 	  {
